Updated archetype:generate command for correctness & clarity.
Re-added archetype details (groupId, artifactId)

Changed sample values for groupId, artifactId, package to match those
suggested earlier in the sample text;  This makes it clearer what's
config to use the plugin, and what's config for the user's project.

The artifacts don't seem to be present in the Maven repository ATM, I've
asked Ross to take a look.